+++ changelog.md
@@ -1,1 +1,2 @@
-- Improved 3D audio
+- Use SocketAddress instead of InetAddress and port
+- Set IPTOS_RELIABILITY traffic class